+++ tests/test_timecorr.py
@@ -41,33 +41,33 @@ class TestTimecorr(unittest.TestCase):
         t = np.array([1, 2, 3, 4, 5, 6], dtype=np.float32)
         r = np.array([2, 3, 4, 5, 6, 7], dtype=np.float32)
         result = timecorr.correct_time(t, r, order)
-        self.assertTrue(np.isclose(result["t_corr"], r).all())
+        self.assertTrue(np.isclose(result.t_corr, r).all())
 
         t = np.array([-2, -1, 0, 1, 2, 3], dtype=np.float32)
         r = np.array([1, 2, 3, 4, 5, 6], dtype=np.float32)
         result = timecorr.correct_time(t, r, order)
-        self.assertTrue(np.isclose(result["t_corr"], r).all())
+        self.assertTrue(np.isclose(result.t_corr, r).all())
 
         # changed order
         order = 3
         t = np.array([1, 2, 3, 4, 5, 6], dtype=np.float32)
         r = np.array([2, 3, 4, 5, 6, 7], dtype=np.float32)
         result = timecorr.correct_time(t, r, order)
-        self.assertTrue(np.isclose(result["t_corr"], r).all())
+        self.assertTrue(np.isclose(result.t_corr, r).all())
 
         # inclination only
         order = 1
         t = np.array([1, 2, 3, 4, 5, 6], dtype=np.float32)
         r = np.array([1, 3, 5, 7, 9, 11], dtype=np.float32)
         result = timecorr.correct_time(t, r, order)
-        self.assertTrue(np.isclose(result["t_corr"], r).all())
+        self.assertTrue(np.isclose(result.t_corr, r).all())
 
         # inclination + offset
         order = 1
         t = np.array([1, 2, 3, 4, 5, 6], dtype=np.float32)
         r = np.array([2.0, 3.5, 5.0, 6.5, 8.0, 9.5], dtype=np.float32)
         result = timecorr.correct_time(t, r, order)
-        self.assertTrue(np.isclose(result["t_corr"], r).all())
+        self.assertTrue(np.isclose(result.t_corr, r).all())
 
     def test_pldt_filter(self):
         # edge case: first invalid